+++ router/common.go
@@ -1,0 +1,11 @@
+package router
+
+import "github.com/google/uuid"
+
+type Share struct {
+	Share bool `json:"share"`
+}
+
+type ID struct {
+	ID uuid.UUID `json:"id"`
+}
